+++ Common/src/main/java/at/petrak/hexcasting/datagen/HexAdvancements.java
@@ -58,13 +58,13 @@ public class HexAdvancements extends PaucalAdvancementSubProvider {
             .parent(root)
             .addCriterion("waste_amt", new SpendMediaTrigger.Instance(ContextAwarePredicate.ANY,
                 MinMaxBounds.Ints.ANY,
-                MinMaxBounds.Ints.atLeast(89 * MediaConstants.DUST_UNIT / 10)))
+                MinMaxBounds.Ints.atLeast((int) (89 * MediaConstants.DUST_UNIT / 10))))
             .save(consumer, prefix("aaa_wasteful_cast"));
         Advancement.Builder.advancement()
             .display(simpleDisplay(HexItems.CHARGED_AMETHYST, "big_cast", FrameType.TASK))
             .parent(root)
             .addCriterion("cast_amt", new SpendMediaTrigger.Instance(ContextAwarePredicate.ANY,
-                MinMaxBounds.Ints.atLeast(64 * MediaConstants.CRYSTAL_UNIT),
+                MinMaxBounds.Ints.atLeast((int) (64 * MediaConstants.CRYSTAL_UNIT)),
                 MinMaxBounds.Ints.ANY))
             .save(consumer, prefix("aab_big_cast"));
